+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # to be released...
 
+- **BC Break:** `Dashboard` attributes `sideMenu`, `topRightMenu` and `topLeftMenu` are now **deprecated**. They are moved to the `stores.*` property.
+- Add `Dashboard.stores` property, to allow dynamic configuration of the dashboard, especially when updating menus _after_ the dashboard is loaded.
 - Add `htmlAttributes` option to `Action` url/button types.
 
 # v0.13.0